Refactored to remove open and close emoji from context
